python/legacy/services/zk-service/pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "zk-service"
version = "0.1.0"
description = "Core services / OMS / APIs"
requires-python = ">=3.10"
dependencies = [
    "zk-core>=0.1.0",
    "zk-proto-betterproto>=0.1.0",
    "zk-rpc>=0.1.0",
    "zk-client>=0.1.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "redis>=5.0",
    "nats-py>=2.6",
    "loguru>=0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[project.scripts]
zk-oms = "tq_service_oms.oms_server:main"
zk-strat = "tq_service_strategy.strat_service:main"

[tool.hatch.build]
include = ["src/**"]

[tool.hatch.build.targets.wheel]
packages = [
    "src/tq_service_api",
    "src/tq_service_app",
    "src/tq_service_helper",
    "src/tq_service_oms",
    "src/tq_service_riskmonitor",
    "src/tq_service_simulator",
    "src/tq_service_strategy",
]
